fix: satisfy audit for finding contract
